Move TextTranslation and StringTranslation under KeyValue namespace
