+++ lib/mobility/backends/sequel/key_value.rb
@@ -4,8 +4,6 @@ require "mobility/backends/sequel"
 require "mobility/backends/key_value"
 require "mobility/sequel/column_changes"
 require "mobility/sequel/hash_initializer"
-require "mobility/sequel/string_translation"
-require "mobility/sequel/text_translation"
 
 module Mobility
   module Backends
@@ -33,7 +31,7 @@ Implements the {Mobility::Backends::KeyValue} backend for Sequel models.
           super
           if type = options[:type]
             options[:association_name] ||= :"#{options[:type]}_translations"
-            options[:class_name]       ||= Mobility::Sequel.const_get("#{type.capitalize}Translation")
+            options[:class_name]       ||= const_get("#{type.capitalize}Translation")
           end
           options[:table_alias_affix] = "#{model_class}_%s_#{options[:association_name]}"
         rescue NameError
@@ -163,7 +161,7 @@ Implements the {Mobility::Backends::KeyValue} backend for Sequel models.
 
       # Returns translation for a given locale, or initializes one if none is present.
       # @param [Symbol] locale
-      # @return [Mobility::Sequel::TextTranslation,Mobility::Sequel::StringTranslation]
+      # @return [Mobility::Backends::Sequel::KeyValue::TextTranslation,Mobility::Backends::Sequel::KeyValue::StringTranslation]
       def translation_for(locale, **)
         translation = model.send(association_name).find { |t| t.key == attribute && t.locale == locale.to_s }
         translation ||= class_name.new(locale: locale, key: attribute)
@@ -183,7 +181,7 @@ Implements the {Mobility::Backends::KeyValue} backend for Sequel models.
         def after_destroy
           super
           [:string, :text].freeze.each do |type|
-            Mobility::Sequel.const_get("#{type.capitalize}Translation").
+            Mobility::Backends::Sequel::KeyValue.const_get("#{type.capitalize}Translation").
               where(translatable_id: id, translatable_type: self.class.name).destroy
           end
         end
@@ -210,6 +208,64 @@ Implements the {Mobility::Backends::KeyValue} backend for Sequel models.
           @attribute_name = attribute_name || column
           super(table, column)
         end
+      end
+
+      module Translation
+        def self.included(base)
+          base.class_eval do
+            plugin :validation_helpers
+
+            # Paraphased from sequel_polymorphic gem
+            #
+            model = underscore(self.to_s)
+            plural_model = pluralize(model)
+            many_to_one :translatable,
+              reciprocal: plural_model.to_sym,
+              reciprocal_type: :many_to_one,
+              setter: (proc do |able_instance|
+                self[:translatable_id]   = (able_instance.pk if able_instance)
+                self[:translatable_type] = (able_instance.class.name if able_instance)
+              end),
+              dataset: (proc do
+                translatable_type = send :translatable_type
+                translatable_id   = send :translatable_id
+                return if translatable_type.nil? || translatable_id.nil?
+                klass = self.class.send(:constantize, translatable_type)
+                klass.where(klass.primary_key => translatable_id)
+              end),
+              eager_loader: (proc do |eo|
+                id_map = {}
+                eo[:rows].each do |model|
+                  model_able_type = model.send :translatable_type
+                  model_able_id = model.send :translatable_id
+                  model.associations[:translatable] = nil
+                  ((id_map[model_able_type] ||= {})[model_able_id] ||= []) << model if !model_able_type.nil? && !model_able_id.nil?
+                end
+                id_map.each do |klass_name, id_map|
+                  klass = constantize(camelize(klass_name))
+                  klass.where(klass.primary_key=>id_map.keys).all do |related_obj|
+                    id_map[related_obj.pk].each do |model|
+                      model.associations[:translatable] = related_obj
+                    end
+                  end
+                end
+              end)
+
+            def validate
+              super
+              validates_presence [:locale, :key, :translatable_id, :translatable_type]
+              validates_unique   [:locale, :key, :translatable_id, :translatable_type]
+            end
+          end
+        end
+      end
+
+      class TextTranslation < ::Sequel::Model(:mobility_text_translations)
+        include Translation
+      end
+
+      class StringTranslation < ::Sequel::Model(:mobility_string_translations)
+        include Translation
       end
     end
