--- lib/mobility/sequel/text_translation.rb
@@ -1,10 +1,0 @@
-# frozen_string_literal: true
-require "mobility/sequel/translation"
-
-module Mobility
-  module Sequel
-    class TextTranslation < ::Sequel::Model(:mobility_text_translations)
-      include Translation
-    end
-  end
-end
